Refactor shifter
